+++ NEsper/NEsper/compat/threading/IReaderWriterLockCommon.cs
@@ -5,17 +5,17 @@
 	/// </summary>
 	public interface IReaderWriterLockCommon
 	{
-		/// <summary>
-        /// Acquires the reader lock.
-        /// </summary>
-        /// <param name="timeout">The timeout.</param>
-        void AcquireReaderLock(int timeout);
+	    /// <summary>
+	    /// Acquires the reader lock.
+	    /// </summary>
+	    /// <param name="timeout">The timeout.</param>
+	    void AcquireReaderLock(long timeout);
 
-        /// <summary>
-        /// Acquires the writer lock.
-        /// </summary>
-        /// <param name="timeout">The timeout.</param>
-        void AcquireWriterLock(int timeout);
+	    /// <summary>
+	    /// Acquires the writer lock.
+	    /// </summary>
+	    /// <param name="timeout">The timeout.</param>
+	    void AcquireWriterLock(long timeout);
 
         /// <summary>
         /// Releases the reader lock.
@@ -30,17 +30,17 @@
 	
 	public interface IReaderWriterLockCommon<T>
 	{
-		/// <summary>
-        /// Acquires the reader lock.
-        /// </summary>
-        /// <param name="timeout">The timeout.</param>
-        T AcquireReaderLock(int timeout);
+	    /// <summary>
+	    /// Acquires the reader lock.
+	    /// </summary>
+	    /// <param name="timeout">The timeout.</param>
+	    T AcquireReaderLock(long timeout);
 
-        /// <summary>
-        /// Acquires the writer lock.
-        /// </summary>
-        /// <param name="timeout">The timeout.</param>
-        T AcquireWriterLock(int timeout);
+	    /// <summary>
+	    /// Acquires the writer lock.
+	    /// </summary>
+	    /// <param name="timeout">The timeout.</param>
+	    T AcquireWriterLock(long timeout);
 
         /// <summary>
         /// Releases the reader lock.
