+++ NEsper/NEsper/compat/threading/MonitorSlimLock.cs
@@ -102,9 +102,9 @@ namespace com.espertech.esper.compat.threading
             return new TrackedDisposable(InternalRelease);
         }
 
-        public IDisposable Acquire(int msec)
+        public IDisposable Acquire(long msec)
         {
-            InternalAcquire(msec);
+            InternalAcquire((int) msec);
             return new TrackedDisposable(InternalRelease);
         }
 
@@ -116,9 +116,9 @@ namespace com.espertech.esper.compat.threading
         /// <param name="releaseLock"></param>
         /// <param name="msec"></param>
         /// <returns></returns>
-        public IDisposable Acquire(bool releaseLock, int? msec = null)
+        public IDisposable Acquire(bool releaseLock, long? msec = null)
         {
-            InternalAcquire(msec ?? _uLockTimeout);
+            InternalAcquire((int)(msec ?? _uLockTimeout));
             if (releaseLock)
                 return new TrackedDisposable(InternalRelease);
             return new VoidDisposable();
