+++ NEsper/NEsper/compat/threading/StandardReaderWriterLock.cs
@@ -51,9 +51,9 @@ namespace com.espertech.esper.compat.threading
 
 #if DEBUG
         /// <summary>
-        /// Gets or sets a value indicating whether this <see cref="StandardReaderWriterLock"/> is trace.
+        /// Gets or sets a value indicating whether this <see cref="StandardReaderWriterLock"/> is TRACE.
         /// </summary>
-        /// <value><c>true</c> if trace; otherwise, <c>false</c>.</value>
+        /// <value><c>true</c> if TRACE; otherwise, <c>false</c>.</value>
         public bool Trace { get; set; }
 #endif
 
@@ -66,11 +66,11 @@ namespace com.espertech.esper.compat.threading
         /// Acquires the reader lock.
         /// </summary>
         /// <param name="timeout">The timeout.</param>
-        public void AcquireReaderLock(int timeout)
+        public void AcquireReaderLock(long timeout)
         {
             try
             {
-                _rwLock.AcquireReaderLock(timeout);
+                _rwLock.AcquireReaderLock((int) timeout);
             }
             catch(ApplicationException)
             {
@@ -82,11 +82,11 @@ namespace com.espertech.esper.compat.threading
         /// Acquires the writer lock.
         /// </summary>
         /// <param name="timeout">The timeout.</param>
-        public void AcquireWriterLock(int timeout)
+        public void AcquireWriterLock(long timeout)
         {
             try
             {
-                _rwLock.AcquireWriterLock(timeout);
+                _rwLock.AcquireWriterLock((int) timeout);
             }
             catch(ApplicationException)
             {
